Add file structure
Segment app structure into navigation, screens and components.
